Use ctest's BUILD_TESTING option, not our own
CTest already defines a BUILD_TESTING option defining whether tests
added through add_test should be generated (and thus whether a
high level 'test' target should exist) or not. We thus don't need to
define our own BUILD_TESTS option. The flag is ON by default.

Care must still be taken to not *build* tests when the flag is OFF, and
thus we are still conditionally adding more targets to the build system.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>