Adjust references to old source code paths
The path where a project's repository is cloned into is the not same
from runner to runner (or even within the same runner, for different
jobs). This makes source code path references in exported artifacts
invalid, and the Makefiles generated by cmake, and the
compile_commands.json file, are full of those.

This commit tries to overcome that limitation by recording the source
code location of jobs exporting artifacts, and then using that
information to patch the artifacts after they are downloaded by
subsequent jobs. An alternative method was seek where we set the
GIT_CLONE_PATH variable to a fixed value within $CI_BUILDS_DIR, but that
depends on a particular configuration setting on the executors being
turned on (it was off in Nexus'), and on $CI_BUILDS_DIR having the same
value across runners, which is not necessarily true either.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>