Skip binaries when replacing source dir references
Jobs running in docker executors might not run inside the same directory
on the filesystem, as different jobs might be given different working
directories. We previously adjusted our GitLab CI pipeline so that the
artifacts in the build/ directory that are passed from one job to
another would contains references to the correct working directory after
being retrieved (e.g., all the absolute paths contained in the Makefiles
generated by cmake). This took some of the problems away from us, but
apparently is causing *binary* files that might contain this information
embedded on them to fail loading.

This commit changes the way we amend the this absolute path in files,
basically skipping any executable, as well as known binary files like
shared and static libraries, and object files.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>